sharding database&table demo commit

+++ Sharding-JDBC-demo/src/main/java/github/weizibin/algorithm/PreciseModuloTableShardingAlgorithm.java
@@ -5,10 +5,10 @@ import io.shardingjdbc.core.api.algorithm.sharding.standard.PreciseShardingAlgor
 
 import java.util.Collection;
 
-public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Integer> {
+public final class PreciseModuloTableShardingAlgorithm implements PreciseShardingAlgorithm<Long> {
     
     @Override
-    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Integer> shardingValue) {
+    public String doSharding(final Collection<String> availableTargetNames, final PreciseShardingValue<Long> shardingValue) {
         for (String each : availableTargetNames) {
             if (each.endsWith(shardingValue.getValue() / 2 % 2 + "")) {
                 return each;
